move memsys tests to test_nrt_refct
